// frontend/src/lib/types.ts
export interface Class {
  id: string;
  name: string;
  level: number;
  created_at: string;
}
// ...
export interface ClassCocurricularConfig {
  id: string;
  class_id: string;
  has_cocurricular: boolean;
  created_at: string;
}